src/api/requirements.txt
# Dependências mínimas para o pipeline CI funcionar

# Framework principal
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
# pyarrow==10.0.0
# fastparquet==2024.2.0
# pandas==2.3.2
pandas==2.2.2
fastparquet==2024.2.0
pyarrow==21.0.0
openpyxl==3.1.2
numpy<2.0
google-cloud-storage==3.4.0
google-auth==2.26.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
# Ferramentas de qualidade de código e testes
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2
ruff==0.1.6
mypy==1.7.1

# Dependências básicas para testes (Altere conforme necessário)
requests==2.31.0
types-requests==2.31.0.10

